Fix underscore in numbers and remove enum and type
Underscores were not explicitly handled and that was causing the
implicit cast to floating point numbers to truncate at the first
underscore. Now, the scanner ignores them because the proper behavior
for underscores in numbers is not defined so I need to ask Dr. Wilsey
for further clarification.

The enum and type reserved words were removed to comply with the spec
changes. I also renamed token_type to TokenType to align with my style
preferences better.